--- app/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Sep  9 17:11:58 2021
-
-@author: link4
-"""
