--- app/__main__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Sep  9 17:08:07 2021
-
-@author: link4
-"""
-
-if __name__ == '__main__':
-    from server import app
-    app.run()
