+++ requirements.txt
@@ -1,3 +1,5 @@
 flask==1.1.2
 pandas==1.2.1
-tensorflow==2.4.1
+pickle==4.0
+sklearn==0.24.1
+gunicorn==20.1.0
