+++ wsgi.py
@@ -5,7 +5,7 @@ Created on Thu Sep  9 17:01:11 2021
 @author: link4
 """
 
-from app.server import app
+from server import app
 application = app
 
 if __name__ == '__main__':
